+++ src/vs/workbench/contrib/terminal/browser/terminalService.ts
@@ -296,9 +296,7 @@ export class TerminalService implements ITerminalService {
 			this._onDidFocusInstance.fire(instance);
 			this._evaluateActiveInstance(host, instance);
 		});
-		host.onDidChangeInstanceCapability((instance) => {
-			this._onDidChangeInstanceCapability.fire(instance);
-		});
+		host.onDidChangeInstanceCapability(this._onDidChangeInstanceCapability.fire, this.onDidChangeInstanceCapability);
 		this._hostActiveTerminals.set(host, undefined);
 	}
