+++ src/vs/workbench/contrib/terminal/browser/terminalView.ts
@@ -45,7 +45,6 @@ import { withNullAsUndefined } from 'vs/base/common/types';
 import { DataTransfers } from 'vs/base/browser/dnd';
 import { getTerminalActionBarArgs } from 'vs/workbench/contrib/terminal/browser/terminalMenus';
 import { TerminalContextKeys } from 'vs/workbench/contrib/terminal/common/terminalContextKey';
-import { getShellIntegrationTooltip } from 'vs/workbench/contrib/terminal/browser/terminalTooltip';
 
 export class TerminalViewPane extends ViewPane {
 	private _actions: IAction[] | undefined;
@@ -392,10 +391,6 @@ class SingleTerminalTabActionViewItem extends MenuEntryActionViewItem {
 				this.updateLabel();
 			}
 		}));
-		this._register(this._terminalService.onDidChangeInstanceCapability(e => {
-			this._action.tooltip = getSingleTabTooltip(e, this._terminalService.configHelper.config.tabs.separator);
-			this.updateLabel(e);
-		}));
 
 		// Clean up on dispose
 		this._register(toDisposable(() => dispose(this._elementDisposables)));
@@ -509,7 +504,7 @@ function getSingleTabLabel(instance: ITerminalInstance | undefined, separator: s
 		return '';
 	}
 	let iconClass = ThemeIcon.isThemeIcon(instance.icon) ? instance.icon?.id : Codicon.terminal.id;
-	const label = `$(${icon?.id || iconClass}) ${getSingleTabTitle(instance, separator)}`;
+	const label = `$(${icon?.id || iconClass}) ${getSingleTabTooltip(instance, separator)}`;
 
 	const primaryStatus = instance.statusList.primary;
 	if (!primaryStatus?.icon) {
@@ -522,16 +517,10 @@ function getSingleTabTooltip(instance: ITerminalInstance | undefined, separator:
 	if (!instance) {
 		return '';
 	}
-	const shellIntegrationString = getShellIntegrationTooltip(instance);
-	const title = getSingleTabTitle(instance, separator);
-	return shellIntegrationString ? title + shellIntegrationString : title;
-}
-
-function getSingleTabTitle(instance: ITerminalInstance | undefined, separator: string): string {
-	if (!instance) {
-		return '';
+	if (!instance.description) {
+		return instance.title;
 	}
-	return !instance.description ? instance.title : `${instance.title} ${separator} ${instance.description}`;
+	return `${instance.title} ${separator} ${instance.description}`;
 }
 
 class TerminalThemeIconStyle extends Themable {
